Make corrections
