+++ queries.sql
@@ -1,9 +1,9 @@
 /* Добавляем данные в таблицу content_types */
 INSERT INTO content_types (id, name, icon_class) VALUES
-(1, 'Текст', 'text'),
-(2, 'Цитата', 'quote'),
-(3, 'Картинка', 'photo'),
-(4, 'Видео', 'video'),
+(1, 'Фото', 'photo'),
+(2, 'Видео', 'video'),
+(3, 'Текст', 'text'),
+(4, 'Цитата', 'quote'),
 (5, 'Ссылка', 'link');
 
 /* Добавляем данные в таблицу users */
@@ -14,8 +14,8 @@ INSERT INTO users (id, registration_date, email, login, password, avatar) VALUES
 
 /* Добавляем данные в таблицу posts */
 INSERT INTO posts (id, created_at, title, content, quote_author, image, video, link, views, author_id, content_type_id) VALUES
-(1, '2021-06-16 12:00:05', 'Цитата', 'Мы в жизни любим только раз, а после ищем лишь похожих', '', '', '', '', 5, 2, 2),
-(2, '2021-06-15 15:14:21', 'Наконец, обработал фотки!', '', '', 'rock-medium.jpg', '', '', 10, 1, 3),
+(1, '2021-06-16 12:00:05', 'Цитата', 'Мы в жизни любим только раз, а после ищем лишь похожих', '', '', '', '', 5, 2, 4),
+(2, '2021-06-15 15:14:21', 'Наконец, обработал фотки!', '', '', 'rock-medium.jpg', '', '', 10, 1, 1),
 (3, '2021-06-17 11:15:12', 'Лучшие курсы', '', '', '', '', 'www.htmlacademy.ru', 15, 3, 5);
 
 /* Добавляем данные в таблицу comments */
